+++ axelor-production/src/main/java/com/axelor/apps/production/module/ProductionModule.java
@@ -39,6 +39,10 @@ import com.axelor.apps.production.service.BillOfMaterialService;
 import com.axelor.apps.production.service.BillOfMaterialServiceImpl;
 import com.axelor.apps.production.service.MrpForecastProductionService;
 import com.axelor.apps.production.service.MrpForecastProductionServiceImpl;
+import com.axelor.apps.production.service.MpsChargeService;
+import com.axelor.apps.production.service.MpsChargeServiceImpl;
+import com.axelor.apps.production.service.MpsWeeklyScheduleService;
+import com.axelor.apps.production.service.MpsWeeklyScheduleServiceImpl;
 import com.axelor.apps.production.service.MrpLineServiceProductionImpl;
 import com.axelor.apps.production.service.MrpServiceProductionImpl;
 import com.axelor.apps.production.service.ProdProcessLineService;
@@ -135,5 +139,7 @@ public class ProductionModule extends AxelorModule {
         .to(ConfiguratorCreatorImportServiceProductionImpl.class);
     bind(ProductStockLocationServiceImpl.class).to(ProductionProductStockLocationServiceImpl.class);
     bind(MrpForecastProductionService.class).to(MrpForecastProductionServiceImpl.class);
+    bind(MpsWeeklyScheduleService.class).to(MpsWeeklyScheduleServiceImpl.class);
+    bind(MpsChargeService.class).to(MpsChargeServiceImpl.class);
   }
 }