+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 
 ## [Unreleased 6.0]
 ## Features
+
 - Production : Remove stockLocation from Machine Type object.
 - Production : Manage MPS (Master Production Schedule) process.
 - MANUFACTURING : New object S&OP (PIC)
 - Partner : Added a new partner type 'Subcontractor' and added field related to outsourcing in manufacturing
+- PRODUCTION : Creation of a new object MPS Charge
+
 
 ## Improvements
 - Production : Change in reference of machineWorkCenter from workCenter to machine.
@@ -14,3 +17,5 @@
 - Production : Fix Java Heap Error.
 - DEMO DATA : fix issue in BudgetService.
 
+
+[Unreleased 6.0]: https://github.com/axelor/axelor-business-suite/compare/dev...wip